refactor(attestation-service): give the wire-error contract its own module
Fixes SEI-281

The four JSON-RPC error codes this service answers with are a
cross-process contract: a joining node reads the code and decides from
it alone whether its own evidence was refused or whether this responder
merely could not answer. They lived in utils.rs, next to is_sudo()
shelling out to id -u and the tracing setup.

Move the codes, their constructors, and the AdmissionDenial routing into
rpc_error.rs, with the two root-key codes pub. utils.rs keeps the
process-level helpers. The module doc states the contract the codes
serve, including why every message is a stable constant: an
unauthenticated caller learns the outcome and nothing about why, so the
detail stays in this responder's log.

Within the module the standard codes come first (invalid params, internal
error), then the root-key contract and the routing that chooses between
its two codes.

The admission suite asserted against hand-copied literals of both codes
and imports them now, so each code has one definition.

A pure move otherwise: same codes, same messages, same routing.
